requirements-doc.txt
sphinx>=1.8.1,<2
sphinx-autodoc-typehints>=1.3.0,<2
sphinx-rtd-theme>=0.4.2,<1
sphinx-icontract>=2.0.0,<3